+++ settings/webpack.production.js
@@ -3,7 +3,7 @@ const LodashModuleReplacementPlugin = require('lodash-webpack-plugin');
 const TerserPlugin = require('terser-webpack-plugin');
 const CompressionPlugin = require('compression-webpack-plugin');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
-
+const { resolve } = require('path');
 const common = require('./webpack.common.js');
 
 process.env.NODE_ENV = 'production';